refactor(scan): move scan job bookkeeping out of the socket endpoint
Finding a scan in flight and pruning ones left behind is RQ bookkeeping,
not socket work, and the watcher and the startup script both ask for it.
It now lives in a handler that depends on nothing heavier than the queue,
with its own tests; the socket module keeps only what it says to a client.

The func name a directly enqueued scan carries is spelled out rather than
read off the function, which lives in the module that imports this one. A
test holds the two together.
