+++ backend/handler/scan_jobs.py
@@ -1,0 +1,134 @@
+"""Finding and pruning the RQ jobs that run a library scan.
+
+Kept out of the socket endpoint so the watcher and the startup script can ask
+about scans without importing the scan implementation and everything under it.
+"""
+
+from datetime import datetime, timedelta, timezone
+from itertools import chain
+from typing import Final
+
+from rq import Worker
+from rq.job import Job, JobStatus
+from rq.registry import ScheduledJobRegistry
+
+from handler.redis_handler import (
+    get_job_func_name,
+    get_job_status,
+    get_worker_current_job,
+    high_prio_queue,
+    low_prio_queue,
+    redis_client,
+)
+from logger.logger import log
+from tasks.tasks import TaskType
+
+# The name RQ records for a directly enqueued scan. Spelled out rather than
+# derived from the function, which lives in the module that imports this one;
+# a test holds the two together.
+SCAN_PLATFORMS_FUNC: Final = "endpoints.sockets.scan.scan_platforms"
+
+# A delayed watcher scan this far past due was left behind by an instance that
+# was not running, and the change it reacted to has long since settled.
+STALE_SCHEDULED_SCAN_AGE: Final = timedelta(hours=1)
+
+
+def is_scan_job(job: Job) -> bool:
+    """Whether this job runs a scan.
+
+    Socket and watcher scans enqueue scan_platforms itself. Task-driven scans go
+    through the task runner, which every task shares, so they are recognised by
+    the type their job carries instead.
+    """
+    if get_job_func_name(job) == SCAN_PLATFORMS_FUNC:
+        return True
+
+    return job.meta.get("task_type") == TaskType.SCAN
+
+
+def get_running_scan_job() -> Job | None:
+    """The scan currently executing on a worker, if any.
+
+    A started job is no longer in the queue, so it can only be found by asking
+    the workers what they are holding.
+    """
+    for worker in Worker.all(connection=redis_client):
+        job = get_worker_current_job(worker)
+        if job is not None and is_scan_job(job):
+            return job
+
+    return None
+
+
+def get_queued_scan_jobs() -> list[Job]:
+    """Scans sitting on a worker queue, waiting to be picked up.
+
+    Socket scans go to the high priority queue and watcher scans to the low
+    priority one, once their delay is up.
+    """
+    job_ids = chain(high_prio_queue.get_job_ids(), low_prio_queue.get_job_ids())
+    jobs = Job.fetch_many(job_ids, connection=redis_client)
+
+    return [
+        job
+        for job in jobs
+        if job is not None and is_scan_job(job)
+        # The fetch above already carries the status, so re-reading it would be
+        # a round trip per queued job.
+        and get_job_status(job, refresh=False) == JobStatus.QUEUED
+    ]
+
+
+def _scheduled_scan_registry() -> ScheduledJobRegistry:
+    """Where delayed scans wait, which is only ever the watcher's queue."""
+    return ScheduledJobRegistry(queue=low_prio_queue)
+
+
+def get_scheduled_scan_jobs() -> list[Job]:
+    """Scans waiting out a delay, which only the watcher sets.
+
+    These never stand in for a scan in flight: a worker has to be running to
+    release them, so counting them would refuse scans on an idle instance.
+    """
+    registry = _scheduled_scan_registry()
+    jobs = Job.fetch_many(registry.get_job_ids(), connection=redis_client)
+
+    return [job for job in jobs if job is not None and is_scan_job(job)]
+
+
+def get_pending_scan_jobs() -> list[Job]:
+    """Scans that have not started yet: queued, or waiting out a delay.
+
+    A scan already running is deliberately not one of these. It may have walked
+    past the folder that just changed, so a fresh scan is still warranted.
+    """
+    return get_queued_scan_jobs() + get_scheduled_scan_jobs()
+
+
+def drop_stale_scheduled_scans() -> int:
+    """Drop delayed watcher scans that are long past due.
+
+    Releasing a backlog of them at once, which is what an instance that was down
+    for a while does on start, would run the same library scan over and over.
+
+    Returns:
+        int: How many scans were dropped.
+    """
+    registry = _scheduled_scan_registry()
+    cutoff = datetime.now(timezone.utc) - STALE_SCHEDULED_SCAN_AGE
+
+    # The registry is scored by due time, so it can hand back only what is past
+    # due rather than every delayed job.
+    stale_ids = registry.get_jobs_to_schedule(int(cutoff.timestamp()))
+    jobs = Job.fetch_many(stale_ids, connection=redis_client)
+    dropped = 0
+
+    for job in jobs:
+        if job is None or not is_scan_job(job):
+            continue
+
+        job.cancel()
+        dropped += 1
+        log.warning(f"Dropped scan {job.id}, due over {STALE_SCHEDULED_SCAN_AGE} ago")
+
+    return dropped
