+++ backend/watcher.py
@@ -19,11 +19,7 @@ from config import (
     TASK_RESULT_TTL,
 )
 from config.config_manager import config_manager as cm
-from endpoints.sockets.scan import (
-    get_pending_scan_jobs,
-    scan_job_meta,
-    scan_platforms,
-)
+from endpoints.sockets.scan import scan_job_meta, scan_platforms
 from handler.database import db_platform_handler
 from handler.metadata import (
     meta_csdb_handler,
@@ -45,6 +41,7 @@ from handler.metadata import (
 )
 from handler.redis_handler import get_job_kwargs, low_prio_queue
 from handler.scan_handler import MetadataSource, ScanType
+from handler.scan_jobs import get_pending_scan_jobs
 from logger.formatter import CYAN
 from logger.formatter import highlight as hl
 from logger.logger import log
